apps/shorts-generator/src/shorts_generator/services/audio_generator.py
"""Audio Generation Service using Microsoft Edge-TTS.

This service generates audio for short video narration:
- Text-to-speech using Edge-TTS (free)
- Background music mixing
- Closed captions generation in SRT format

Cost: FREE
"""

import logging
import math
import os
import tempfile
from dataclasses import dataclass
from datetime import timedelta
from pathlib import Path
from typing import Any

import httpx

from shorts_generator.core.config import settings

logger = logging.getLogger(__name__)

# Edge-TTS API endpoint (uses undocumented API)
EDGE_TTS_API = "https://speech.platform.bing.com/consumer/speech/synthesize/readaloud/edge/v1"


@dataclass
class GeneratedAudio:
    """Represents generated audio.

    Attributes:
        url: URL to the audio file
        duration_seconds: Audio duration in seconds
        file_path: Local file path (for temporary files)
        generation_method: How the audio was generated (edge_tts)
        voice_used: Which voice was used
    """

    url: str
    duration_seconds: float
    file_path: str
    generation_method: str
    voice_used: str


class AudioGenerator:
    """Generates audio for short video narration."""

    # Default audio settings
    SAMPLE_RATE = 24000  # Hz
    BIT_RATE = 128000  # bps

    def __init__(self):
        """Initialize the audio generator."""
        self._http_client: httpx.AsyncClient | None = None

    async def _get_http_client(self) -> httpx.AsyncClient:
        """Get or create HTTP client."""
        if self._http_client is None:
            self._http_client = httpx.AsyncClient(timeout=60.0)
        return self._http_client

    def _estimate_duration(self, text: str, words_per_second: float = 2.5) -> float:
        """Estimate spoken duration from text.

        Args:
            text: Text to estimate duration for
            words_per_second: Average speaking rate (default: 2.5)

        Returns:
            Estimated duration in seconds
        """
        word_count = len(text.split())
        return max(1.0, word_count / words_per_second)

    async def generate_narration(
        self,
        script_text: str,
        voice: str = "en-US-AriaNeural",
    ) -> GeneratedAudio:
        """Generate narration audio using Edge-TTS.

        Args:
            script_text: Text to convert to speech
            voice: Voice to use (default: en-US-AriaNeural)

        Returns:
            GeneratedAudio object with audio URL

        Raises:
            Exception: If TTS generation fails
        """
        logger.info(f"Generating narration with voice: {voice}")

        try:
            client = await self._get_http_client()

            # Build request payload for Edge-TTS
            payload = {
                "text": script_text,
                "voice": voice,
                "outputFormat": "audio-24khz-128kbitrate-mono-mp3",
                "rate": "+0%",  # Normal speed
                "pitch": "+0Hz",  # Normal pitch
                "volume": "+0%",
            }

            # Call Edge-TTS API
            response = await client.post(
                EDGE_TTS_API,
                json=payload,
                headers={
                    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64)",
                },
            )

            if response.status_code != 200:
                error_text = response.text
                logger.error(f"Edge-TTS API error: {error_text}")
                raise Exception(f"Edge-TTS generation failed: {error_text}")

            # Get audio content
            audio_content = response.content

            # Save to temporary file
            with tempfile.NamedTemporaryFile(delete=False, suffix=".mp3") as tmp_file:
                tmp_file.write(audio_content)
                tmp_file.flush()

                # Calculate approximate duration
                duration = self._estimate_duration(script_text)

                audio = GeneratedAudio(
                    url=f"file://{tmp_file.name}",
                    duration_seconds=duration,
                    file_path=tmp_file.name,
                    generation_method="edge_tts",
                    voice_used=voice,
                )

                logger.info(f"Narration generated: {duration:.2f}s, {len(audio_content)} bytes")

                return audio

        except Exception as e:
            logger.error(f"Narration generation failed: {e}")
            raise

    async def add_background_music(
        self,
        narration_url: str,
        music_track: str = "background_lofi.mp3",
        music_volume: float = 0.3,  # 30% volume
    ) -> GeneratedAudio:
        """Add background music to narration.

        Args:
            narration_url: URL or path to narration audio
            music_track: Path or identifier for background music
            music_volume: Music volume (0.0 to 1.0)

        Returns:
            GeneratedAudio with mixed audio

        Raises:
            Exception: If mixing fails or FFmpeg not available
        """
        logger.info(f"Mixing background music at {music_volume*100}% volume")

        try:
            # Import FFmpeg Python
            import ffmpeg

            # Create output file
            with tempfile.NamedTemporaryFile(delete=False, suffix=".mp3") as tmp_file:
                output_path = tmp_file.name

                # Get file paths from URLs
                if narration_url.startswith("file://"):
                    narration_path = narration_url.replace("file://", "")
                else:
                    # Download to temp file
                    narration_path = tempfile.mktemp(suffix=".mp3")
                    # TODO: Implement download from URL

                # Mix audio using FFmpeg
                try:
                    # Get duration of narration
                    probe = ffmpeg.probe(narration_path)
                    narration_duration = float(probe.format["duration"])

                    # Calculate how much music we need (same duration as narration)
                    # In production, you'd loop the music track

                    (
                        ffmpeg
                        .input(narration_path, stream=0, a="apad")
                        .filter_(0, "volume", f"1.0")  # Keep narration at 100%
                    )

                    # For now, just return the narration without music
                    # In production, you'd add: ffmpeg.input(music_track, stream=1, t=loop_duration)
                    # And use: amix=inputs=1, volume=1

                    # Simplified: return narration as-is
                    audio = GeneratedAudio(
                        url=narration_url,
                        duration_seconds=narration_duration,
                        file_path=narration_path,
                        generation_method="edge_tts",  # Still edge_tts since we didn't modify
                        voice_used="en-US-AriaNeural",
                    )

                    return audio

                except ffmpeg.Error as e:
                    logger.error(f"FFmpeg processing failed: {e}")
                    raise Exception(f"Audio mixing failed: {e}")

        except ImportError:
            logger.warning("FFmpeg not available, returning narration without music")
            # Fallback: return narration as-is
            # In production, you'd want to fail gracefully
            return GeneratedAudio(
                url=narration_url,
                duration_seconds=10.0,
                file_path="",
                generation_method="edge_tts",
                voice_used="en-US-AriaNeural",
            )

    def _format_timestamp(self, seconds: float) -> str:
        """Format seconds to SRT timestamp format.

        Args:
            seconds: Time in seconds

        Returns:
            Formatted timestamp (HH:MM:SS,mmm)
        """
        hours = int(seconds // 3600)
        minutes = int((seconds % 3600) // 60)
        secs = seconds % 60
        millis = int((secs % 1) * 1000)

        return f"{hours:02d}:{minutes:02d}:{secs:02d},{millis:03d}"

    async def generate_captions(
        self,
        script_text: str,
        scene_timings: list[dict[str, Any]] | None = None,
    ) -> str:
        """Generate closed captions in SRT format.

        Args:
            script_text: Full script text
            scene_timings: Optional list of scene timings for more accurate captions

        Returns:
            SRT format captions as string
        """
        logger.info("Generating closed captions")

        # Split script into words
        words = script_text.split()
        total_words = len(words)

        # Assume even distribution of words across time
        # In production, you'd use word timings from TTS API if available
        duration_seconds = self._estimate_duration(script_text)
        time_per_word = duration_seconds / total_words

        # Build SRT content
        srt_lines = []

        # SRT format: 1, 00:00:00,000 --> 00:00:02,500
        current_time = 0.0
        caption_index = 1
        caption_words = []

        for i, word in enumerate(words):
            caption_words.append(word)

            # Create a caption every ~10 words or at end
            if len(caption_words) >= 10 or i == total_words - 1:
                caption_text = " ".join(caption_words)
                word_duration = len(caption_words) * time_per_word
                start_time = current_time
                end_time = current_time + word_duration

                srt_lines.append(f"{caption_index}")
                srt_lines.append(f"{self._format_timestamp(start_time)} --> {self._format_timestamp(end_time)}")
                srt_lines.append(caption_text)
                srt_lines.append("")  # Blank line between captions

                current_time = end_time
                caption_index += 1
                caption_words = []

        return "\n".join(srt_lines)

    async def generate_audio_for_script(
        self,
        script_text: str,
        voice: str = "en-US-AriaNeural",
        add_music: bool = False,
    ) -> tuple[GeneratedAudio, str]:
        """Generate complete audio package for a script.

        Args:
            script_text: Script text to narrate
            voice: Voice to use
            add_music: Whether to add background music

        Returns:
            Tuple of (GeneratedAudio, SRT captions)
        """
        # Generate narration
        narration = await self.generate_narration(script_text, voice)

        # Generate captions
        captions = await self.generate_captions(script_text)

        # Add music if requested
        if add_music:
            final_audio = await self.add_background_music(narration.url)
        else:
            final_audio = narration

        return final_audio, captions


# Singleton instance
audio_generator = AudioGenerator()
