apps/shorts-generator/pyproject.toml
[project]
name = "shorts-generator"
version = "0.1.0"
description = "Lesson Shorts Generator - Automated short video generation from lesson content"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "celery>=5.4.0",
    "google-genai>=1.0.0",
    "replicate>=0.31.0",
    "edge-tts>=6.1.0",
    "ffmpeg-python>=0.2.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    "markdown>=3.7.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shorts_generator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
